+++ components/cart/CartContentFooter.js
@@ -1,0 +1,5 @@
+export default function CartContentFooter({ couponList }) {
+  const rateCouponList = couponList.rateCoupon;
+  const amountCouponList = couponList.amountCoupon;
+  return <div>{}</div>;
+}
